+++ playgrounds/demo/app/pages/components/scroll-area.vue
@@ -46,23 +46,25 @@ const virtualizeOptions = computed(() => {
 <template>
   <PlaygroundPage>
     <template #controls>
-      <B24Switch v-model="virtualize" label="Virtualize" />
+      <B24Switch v-model="virtualize" label="Virtualize" size="xs" />
       <B24FieldGroup>
         <B24Button
           active-color="air-primary"
           :active="orientation === 'vertical'"
           :icon="MoreVerticalLIcon"
+          size="xs"
           @click="orientation = 'vertical'"
         />
         <B24Button
           active-color="air-primary"
           :active="orientation === 'horizontal'"
           :icon="MoreLIcon"
+          size="xs"
           @click="orientation = 'horizontal'"
         />
       </B24FieldGroup>
       <template v-if="virtualize">
-        <B24FormField label="gap" orientation="horizontal">
+        <B24FormField label="gap" orientation="horizontal" size="xs">
           <B24InputNumber
             v-model="gap"
             class="w-16"
@@ -70,10 +72,11 @@ const virtualizeOptions = computed(() => {
             :icon="ColumnsIcon"
             :max="48"
             orientation="vertical"
+            size="xs"
           />
         </B24FormField>
 
-        <B24FormField label="lanes" orientation="horizontal">
+        <B24FormField label="lanes" orientation="horizontal" size="xs">
           <B24InputNumber
             v-model="lanes"
             class="w-16"
@@ -81,6 +84,7 @@ const virtualizeOptions = computed(() => {
             :icon="AppsIcon"
             :max="10"
             orientation="vertical"
+            size="xs"
           />
         </B24FormField>
       </template>