+++ playgrounds/nuxt/app/pages/components/avatar.vue
@@ -13,8 +13,15 @@ const showGroup = ref(false)
 <template>
   <PlaygroundPage>
     <template #controls>
-      <B24Select v-model="attrs.size" class="w-32" :items="sizes" placeholder="Size" multiple />
-      <B24Switch v-model="showGroup" label="Group" />
+      <B24Select
+        v-model="attrs.size"
+        class="w-32"
+        :items="sizes"
+        placeholder="Size"
+        multiple
+        size="xs"
+      />
+      <B24Switch v-model="showGroup" label="Group" size="xs" />
     </template>
 
     <Matrix