+++ playgrounds/nuxt/app/pages/components/collapsible.vue
@@ -10,7 +10,7 @@ defineShortcuts({
 <template>
   <PlaygroundPage>
     <template #controls>
-      <B24Switch v-model="disabled" label="Disabled" />
+      <B24Switch v-model="disabled" label="Disabled" size="xs" />
     </template>
 
     <B24Collapsible