+++ playgrounds/nuxt/app/pages/components/pagination.vue
@@ -7,10 +7,10 @@ const showEdges = ref(true)
 <template>
   <PlaygroundPage>
     <template #controls>
-      <B24Switch v-model="showEdges" label="Show edges" />
+      <B24Switch v-model="showEdges" label="Show edges" size="xs" />
 
-      <B24FormField label="Siblings" orientation="horizontal">
-        <B24InputNumber v-model="siblingCount" class="w-20" :min="0" :max="3" />
+      <B24FormField label="Siblings" orientation="horizontal" size="xs">
+        <B24InputNumber v-model="siblingCount" class="w-20" :min="0" :max="3" size="xs" />
       </B24FormField>
     </template>